Implement more of the datasets api

+++ app/Http/Controllers/Web/Datasets/DestroyDatasetWebController.php
@@ -10,6 +10,8 @@ class DestroyDatasetWebController extends Controller
 {
     public function __invoke(Project $project, Dataset $dataset)
     {
+        abort_unless(is_null($dataset->published_at), 400, "Cannot delete a published dataset");
+        abort_unless(is_null($dataset->doi), 400, "Cannot delete a dataset that has a DOI assigned");
         $dataset->delete();
         return redirect(route('projects.datasets.index', compact('project')));
     }
